Added new option to insert new keys if they don’t exist.
I’ve left this off by default as most people will only want to
change an existing key - running with this option enabled will 
mean a typo in the key name will not result in an error. But it
is useful if you need to insert a new key at build time.